+++ Apizr/Apizr.Extensions.Microsoft.DependencyInjection/ApizrExtendedOptions.cs
@@ -16,12 +16,12 @@ namespace Apizr
     public class ApizrExtendedOptions : ApizrOptionsBase, IApizrExtendedOptions
     {
         public ApizrExtendedOptions(Type webApiType, Type apizrManagerType, Uri baseAddress,
-            DecompressionMethods decompressionMethods,
             HttpMessageParts? httpTracerVerbosity, string[] assemblyPolicyRegistryKeys,
-            string[] webApiPolicyRegistryKeys) : base(webApiType, baseAddress, decompressionMethods,
+            string[] webApiPolicyRegistryKeys) : base(webApiType, baseAddress,
             httpTracerVerbosity, assemblyPolicyRegistryKeys, webApiPolicyRegistryKeys)
         {
             ApizrManagerType = apizrManagerType;
+            HttpClientHandlerFactory = _ => new HttpClientHandler();
             RefitSettingsFactory = _ => new RefitSettings();
             ConnectivityHandlerType = typeof(VoidConnectivityHandler);
             CacheHandlerType = typeof(VoidCacheHandler);
@@ -39,6 +39,7 @@ namespace Apizr
         public Type CacheHandlerType { get; set; }
         public Type LogHandlerType { get; set; }
         public Type MappingHandlerType { get; set; }
+        public Func<IServiceProvider, HttpClientHandler> HttpClientHandlerFactory { get; set; }
         public Func<IServiceProvider, RefitSettings> RefitSettingsFactory { get; set; }
         public Action<IHttpClientBuilder> HttpClientBuilder { get; set; }
         public IList<Func<IServiceProvider, DelegatingHandler>> DelegatingHandlersExtendedFactories { get; }