+++ Apizr/Apizr/WebApiAttribute.cs
@@ -13,12 +13,10 @@ namespace Apizr
         /// Define web api <see cref="BaseUri"/> and <see cref="DecompressionMethods"/>
         /// </summary>
         /// <param name="baseUri">The web api base uri</param>
-        /// <param name="decompressionMethods">The web api base decompression methods (default: <see cref="DecompressionMethods.None"/>)</param>
         /// <param name="isAutoRegistrable">Makes decorated interface registrable by assembly scanning (default: true)</param>
-        public WebApiAttribute(string baseUri, DecompressionMethods decompressionMethods = DecompressionMethods.None, bool isAutoRegistrable = true)
+        public WebApiAttribute(string baseUri, bool isAutoRegistrable = true)
         {
             BaseUri = baseUri;
-            DecompressionMethods = decompressionMethods;
             IsAutoRegistrable = isAutoRegistrable;
         }
 
@@ -26,11 +24,6 @@ namespace Apizr
         /// The web api base uri
         /// </summary>
         public string BaseUri { get; }
-
-        /// <summary>
-        /// The web api base decompression methods
-        /// </summary>
-        public DecompressionMethods DecompressionMethods { get; }
 
         /// <summary>
         /// Makes decorated interface registrable by assembly scanning